IDEA-229823: Add more YAML settings to editorconfig
By merging the two CodeStyleSettingsProviders and renaming createSettingsPage to createConfigurable
the editorconfig plugin can now pick up the custom settings that the yaml plugin provides.
